requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
Jinja2==3.1.4
tenacity==9.0.0
notion-client==2.2.1
APScheduler==3.10.4
httpx==0.27.2
python-multipart==0.0.9
PyYAML==6.0.2